Shapes Project
My submission for an OOP class project: I was tasked with developing a
program that would allow the user to create various shapes, of a
specified size and colour, and to move them - individually or together
as a compound shape. For full marks, I was required to have
well-documented code, smooth performance and implement the Observer,
Composite, and Iterator design patterns.